test(ctrip): avoid adapter src import
